DeprecationWarning when calling `salt.utils.cloud.fire_event()` without `sock_dir`
The `sock_dir` parameter should be used when calling `fire_event()`
because it currently defaults to a value that isn't very useful if no
value is passed in. Eventually the default value will be removed, at
which point the `sock_dir` parameter will become a hard requirement. In
the meantime we raise a DeprecationWarning to warn the users of this
impending change.